distributor API
